+++ packages/types/src/api/web/workspace/index.ts
@@ -1,4 +1,3 @@
-export * from "./application"
 export * from "./attachment"
 export * from "./automation"
 export * from "./backup"
@@ -9,10 +8,12 @@ export * from "./features"
 export * from "./integration"
 export * from "./layout"
 export * from "./metadata"
+export * from "./navigation"
 export * from "./oauth2"
 export * from "./permission"
-export * from "./workspaceApp"
 export * from "./query"
+export * from "./recaptcha"
+export * from "./resource"
 export * from "./role"
 export * from "./rowAction"
 export * from "./rows"
@@ -22,7 +23,6 @@ export * from "./table"
 export * from "./user"
 export * from "./view"
 export * from "./webhook"
-export * from "./navigation"
-export * from "./resource"
+export * from "./workspace"
+export * from "./workspaceApp"
 export * from "./workspaceFavourites"
-export * from "./recaptcha"
